requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq==0.13.0
pydantic==2.10.5
python-dotenv==1.0.1
locust==2.32.4
requests==2.32.3